Revert "FE - Advanced Settings Page Validation and Error Handling"
